remove unused pages, naming convention
