+++ client/ionic/src/App.tsx
@@ -3,13 +3,11 @@ import { Redirect, Route } from 'react-router-dom';
 import { IonApp, IonRouterOutlet } from '@ionic/react';
 import { IonReactRouter } from '@ionic/react-router';
 
-import About from './pages/About';
-import CopingEveryone from './pages/CopingEveryone';
-import CopingParents from './pages/CopingParents';
+import About from'./pages/About';
+import CheckYourHealth from './pages/CheckYourHealth';
+import Splash from './pages/Splash';
 import Menu from './pages/Menu';
-import MenuDistress from './pages/MenuDistress';
 import ProtectYourself from './pages/ProtectYourself';
-import Triage from './pages/Triage';
 
 /* Core CSS required for Ionic components to work properly */
 import '@ionic/react/css/core.css';
@@ -38,21 +36,15 @@ const App: React.FC = () => (
     <IonReactRouter>
       <IonRouterOutlet>
         <Route path="/about" component={About} exact={true} />
-        <Route
-          path="/coping-everyone"
-          component={CopingEveryone}
-          exact={true}
-        />
-        <Route path="/coping-parents" component={CopingParents} exact={true} />
+        <Route path="/splash" component={Splash} exact={true} />
+        <Route path="/check-your-health" component={CheckYourHealth} exact={true} />
         <Route path="/menu" component={Menu} exact={true} />
-        <Route path="/menu-distress" component={MenuDistress} exact={true} />
         <Route
           path="/protect-yourself"
           component={ProtectYourself}
           exact={true}
         />
-        <Route path="/triage" component={Triage} />
-        <Route path="/" render={() => <Redirect to="/about" />} exact={true} />
+        <Route path="/" render={() => <Redirect to="/splash" />} exact={true} />
       </IonRouterOutlet>
     </IonReactRouter>
   </IonApp>
